+++ controllers/deletePostController.php
@@ -18,6 +18,15 @@ class deletePostController {
         try {
             if(isset($_POST['envoyer'])) {
 
+                // If token is not difined OR if post token is different from the session token
+                if (!$_POST['token'] || $_POST['token'] !== $_SESSION['TOKEN']) {
+                    // show an error message
+                    echo '<p class="error">Error: invalid form submission</p>';
+                    // return 405 http status code
+                    header($_SERVER['SERVER_PROTOCOL'] . ' 405 Method Not Allowed');
+                    exit;
+                }
+
                 $result = $this->_deletePostRepository->deletePost($id);      
 
                 header("Location: /blog/public/all-posts");
